+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Minesweeper".
 - Tests for "The grand festival - I".
 - Tests for "Depot organization".
+- Tests for "Knights jam".
 
 ## [1.11.0] - 2022-07-31
 ### Added